feat: Add --from-file option to compare_reports.py
I've added a new '--from-file' option to 'compare_reports.py' to allow generating a comparison report from a list of URLs in a file.

I also updated the script to automatically look for the file in the 'url-lists/' directory if a simple filename is provided. I've documented these changes in the README.